Move routers to adapters
